Adjust digging skill up rate
The new rate is not retail either (which is vaguely known), but the consensus among the server responders in the controversial channel we shouldn't allow the rate to be quite so grindy.

This new rate starts at roughly the retail rate but the difficulty increase is linear and much slower in ramping up.

Also, changing skill up roll to 1000 to allow half percents with integers.