Restore pre-Topaz Fields of Valor behavior
Prior to our move to Topaz, a character could earn to Fields of Valor rewards on a given Vana'diel day (i.e., per hour) if he missed the reward for the previous Vana'diel day. But otherwise, it was one FoV reward per day. This small changes restores that functionality by storing the last two times a reward was earned, and doing a potrential early-exit on the reward-granting function based on the data from both.